+++ src/components/NotificationList.tsx
@@ -1,0 +1,29 @@
+import { VStack, Alert, AlertIcon, Box, AlertTitle, CloseButton } from '@chakra-ui/react';
+import React from 'react';
+
+import Notification from '../types/Notification';
+
+interface NotificationProps {
+  notifications: Notification[];
+  setNotifications: React.Dispatch<React.SetStateAction<Notification[]>>;
+}
+
+const NotificationList: React.FC<NotificationProps> = ({ notifications, setNotifications }) => {
+  return (
+    <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
+      {notifications.map((notification, index) => (
+        <Alert key={index} status="info" variant="solid" width="auto">
+          <AlertIcon />
+          <Box flex="1">
+            <AlertTitle fontSize="sm">{notification.message}</AlertTitle>
+          </Box>
+          <CloseButton
+            onClick={() => setNotifications((prev) => prev.filter((_, i) => i !== index))}
+          />
+        </Alert>
+      ))}
+    </VStack>
+  );
+};
+
+export default NotificationList;
