+++ src/components/OverlapAlertDialog.tsx
@@ -1,0 +1,104 @@
+import {
+  AlertDialog,
+  AlertDialogOverlay,
+  AlertDialogContent,
+  AlertDialogHeader,
+  AlertDialogBody,
+  AlertDialogFooter,
+  Button,
+  Text,
+} from '@chakra-ui/react';
+import React, { useRef } from 'react';
+
+import { Event, EventForm } from '../types';
+
+interface OverlapAlertDialogProps {
+  event: Event | EventForm;
+  editingEvent: Event | null;
+  isRepeating: boolean;
+  overlappingEvents: Event[];
+  isOverlapDialogOpen: boolean;
+  setIsOverlapDialogOpen: React.Dispatch<React.SetStateAction<boolean>>;
+
+  saveEvent: (event: Event | EventForm) => void;
+}
+const OverlapAlertDialog: React.FC<OverlapAlertDialogProps> = ({
+  event,
+  editingEvent,
+  isRepeating,
+  isOverlapDialogOpen,
+  setIsOverlapDialogOpen,
+  overlappingEvents,
+  saveEvent,
+}) => {
+  const cancelRef = useRef<HTMLButtonElement>(null);
+  const {
+    title,
+    date,
+    startTime,
+    endTime,
+    description,
+    location,
+    category,
+    repeat,
+    notificationTime,
+  } = event;
+  return (
+    <AlertDialog
+      isOpen={isOverlapDialogOpen}
+      leastDestructiveRef={cancelRef}
+      onClose={() => setIsOverlapDialogOpen(false)}
+    >
+      <AlertDialogOverlay>
+        <AlertDialogContent>
+          <AlertDialogHeader fontSize="lg" fontWeight="bold">
+            일정 겹침 경고
+          </AlertDialogHeader>
+
+          <AlertDialogBody>
+            다음 일정과 겹칩니다:
+            {overlappingEvents.map((event) => (
+              <Text key={event.id}>
+                {event.title} ({event.date} {event.startTime}-{event.endTime})
+              </Text>
+            ))}
+            계속 진행하시겠습니까?
+          </AlertDialogBody>
+
+          <AlertDialogFooter>
+            <Button ref={cancelRef} onClick={() => setIsOverlapDialogOpen(false)}>
+              취소
+            </Button>
+            <Button
+              colorScheme="red"
+              onClick={() => {
+                setIsOverlapDialogOpen(false);
+                saveEvent({
+                  id: editingEvent ? editingEvent.id : undefined,
+                  title,
+                  date,
+                  startTime,
+                  endTime,
+                  description,
+                  location,
+                  category,
+                  repeat: {
+                    type: isRepeating ? repeat.type : 'none',
+                    interval: repeat.interval,
+                    endDate: repeat.endDate || undefined,
+                  },
+                  notificationTime,
+                });
+              }}
+              ml={3}
+            >
+              계속 진행
+            </Button>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialogOverlay>
+    </AlertDialog>
+  );
+};
+
+export default OverlapAlertDialog;
